Fix type for --max-ubo-size and --max-pushconstant-size options (#756)
The corresponding Vulkan limits are uint32_t values and some implementations
report values greater than INT_MAX

See https://github.com/kpet/clvk/issues/318

+++ include/clspv/Option.h
@@ -98,7 +98,7 @@ bool ConstantArgsInUniformBuffer();
 // Returns the maximum UBO size. This size is specified in bytes and is used to
 // calculate the size of UBO arrays for constant arguments if
 // ConstantArgsInUniformBuffer returns true.
-uint64_t MaxUniformBufferSize();
+uint32_t MaxUniformBufferSize();
 
 // Returns the maximum push constant interface size. This size is specified in
 // bytes and is used to validate the the size of the POD kernel interface
